ai-orchestrator/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0
websockets==13.1
httpx==0.27.0
python-dotenv==1.0.0
ollama==0.4.0
langgraph==0.2.28

# RAG & Vector Store (Phase 3)
chromadb==0.5.0
numpy<2.0.0
pypdf==4.2.0
tiktoken==0.7.0

# LLM integration
ollama==0.4.0
google-genai>=1.0.0
pyyaml==6.0.2

# Phase 7: Deep reasoning harness + external MCP
mcp>=1.0.0
anthropic>=0.40.0

